requirements.txt
# Core ML/DL
torch==2.9.1
torchvision==0.24.1
timm==1.0.22
ultralytics==8.3.233

# OCR & Translation
manga-ocr
google-generativeai==0.8.5
transformers==4.57.1
sentencepiece==0.2.1

# Inpainting
simple-lama-inpainting-updated==0.1.3

# Image Processing
opencv-python==4.12.0.88
Pillow==12.0.0
numpy==2.2.6
scipy==1.16.3

# Utilities
tqdm==4.67.1
huggingface-hub==0.36.0
safetensors==0.7.0
einops==0.8.1
PyYAML==6.0.3
matplotlib==3.10.7

# Optional: PaddleOCR
# paddlepaddle-gpu==3.2.2
